move content of sh2drc.cpp back into sh2.cpp rather than #including it, easier to refactor this way (nw)
